[refactor/#224] calendar 폴더명 PascalCase로 변경, import 경로 수정
